requirements.txt
fastapi
easyocr
uvicorn
numpy
Pillow
python-multipart